+++ lib/tornado_py3/process.py
@@ -17,6 +17,7 @@
 the server into multiple processes and managing subprocesses.
 """
 
+import errno
 import os
 import multiprocessing
 import signal
@@ -35,6 +36,7 @@ from tornado_py3 import ioloop
 from tornado_py3.iostream import PipeIOStream
 from tornado_py3.log import gen_log
 from tornado_py3.platform.auto import set_close_exec
+from tornado_py3.util import errno_from_exception
 
 import typing
 from typing import Tuple, Optional, Any, Callable
@@ -55,7 +57,7 @@ def cpu_count() -> int:
     except NotImplementedError:
         pass
     try:
-        return os.sysconf("SC_NPROCESSORS_CONF")  # type: ignore
+        return os.sysconf("SC_NPROCESSORS_CONF")
     except (AttributeError, ValueError):
         pass
     gen_log.error("Could not detect number of processors; assuming 1")
@@ -87,9 +89,7 @@ def _pipe_cloexec() -> Tuple[int, int]:
 _task_id = None
 
 
-def fork_processes(
-    num_processes: Optional[int], max_restarts: Optional[int] = None
-) -> int:
+def fork_processes(num_processes: Optional[int], max_restarts: int = None) -> int:
     """Starts multiple worker processes.
 
     If ``num_processes`` is None or <= 0, we detect the number of cores
@@ -110,17 +110,12 @@ def fork_processes(
     number between 0 and ``num_processes``.  Processes that exit
     abnormally (due to a signal or non-zero exit status) are restarted
     with the same id (up to ``max_restarts`` times).  In the parent
-    process, ``fork_processes`` calls ``sys.exit(0)`` after all child
-    processes have exited normally.
+    process, ``fork_processes`` returns None if all child processes
+    have exited normally, but will otherwise only exit by throwing an
+    exception.
 
     max_restarts defaults to 100.
-
-    Availability: Unix
     """
-    if sys.platform == "win32":
-        # The exact form of this condition matters to mypy; it understands
-        # if but not assert in this context.
-        raise Exception("fork not available on windows")
     if max_restarts is None:
         max_restarts = 100
 
@@ -149,7 +144,12 @@ def fork_processes(
             return id
     num_restarts = 0
     while children:
-        pid, status = os.wait()
+        try:
+            pid, status = os.wait()
+        except OSError as e:
+            if errno_from_exception(e) == errno.EINTR:
+                continue
+            raise
         if pid not in children:
             continue
         id = children.pop(pid)
@@ -273,8 +273,6 @@ class Subprocess(object):
         In many cases a close callback on the stdout or stderr streams
         can be used as an alternative to an exit callback if the
         signal handler is causing a problem.
-
-        Availability: Unix
         """
         self._exit_callback = callback
         Subprocess.initialize()
@@ -296,8 +294,6 @@ class Subprocess(object):
         to suppress this behavior and return the exit status without raising.
 
         .. versionadded:: 4.2
-
-        Availability: Unix
         """
         future = Future()  # type: Future[int]
 
@@ -325,8 +321,6 @@ class Subprocess(object):
         .. versionchanged:: 5.0
            The ``io_loop`` argument (deprecated since version 4.1) has been
            removed.
-
-        Availability: Unix
         """
         if cls._initialized:
             return
@@ -353,9 +347,10 @@ class Subprocess(object):
     @classmethod
     def _try_cleanup_process(cls, pid: int) -> None:
         try:
-            ret_pid, status = os.waitpid(pid, os.WNOHANG)  # type: ignore
-        except ChildProcessError:
-            return
+            ret_pid, status = os.waitpid(pid, os.WNOHANG)
+        except OSError as e:
+            if errno_from_exception(e) == errno.ECHILD:
+                return
         if ret_pid == 0:
             return
         assert ret_pid == pid
@@ -363,14 +358,11 @@ class Subprocess(object):
         subproc.io_loop.add_callback_from_signal(subproc._set_returncode, status)
 
     def _set_returncode(self, status: int) -> None:
-        if sys.platform == "win32":
-            self.returncode = -1
+        if os.WIFSIGNALED(status):
+            self.returncode = -os.WTERMSIG(status)
         else:
-            if os.WIFSIGNALED(status):
-                self.returncode = -os.WTERMSIG(status)
-            else:
-                assert os.WIFEXITED(status)
-                self.returncode = os.WEXITSTATUS(status)
+            assert os.WIFEXITED(status)
+            self.returncode = os.WEXITSTATUS(status)
         # We've taken over wait() duty from the subprocess.Popen
         # object. If we don't inform it of the process's return code,
         # it will log a warning at destruction in python 3.6+.