+++ CHANGES.md
@@ -14,7 +14,7 @@
 * Update PySocks 1.7.0 (91dcdf0) to 1.7.1 (c2fa43c)
 * Update Requests library 2.22.0 (3d968ff) to 2.22.0 (d2f65af)
 * Update Six compatibility library 1.13.0 (ec58185) to 1.14.0 (3a3db75)
-* Update Tornado_py3 Web Server 6.0.3 (ff985fe) to 6.0.xx (18b653c)
+* Update Tornado_py3 Web Server 6.1.dev1 (18b653c) to 6.0.4 (b4e39e5)
 * Update urllib3 release 1.25.6 (4a6c288) to 1.25.7 (37ba61a)
 * Add Telegram notifier
 * Change enable image caching on browse pages
@@ -28,6 +28,7 @@
 * Fix name_parser_tests
 * Fix config/Media Process/Unpack test unrar under py3
 * Change remove deprecated `buildNameCache`
+* Update Tornado_py3 Web Server 6.0.3 (ff985fe) to 6.1.dev1 (18b653c)
 
 
 ### 0.21.18 (2020-03-04 19:20:00 UTC)